# Optimize the spinner rendering