It turns out the spinner animation logic was slowing down Bake in the case where all the tasks are cached and there is no real work to do. I created a bakefile with 30 tasks in it. Previously, Bake would take 3.04 seconds to go through all those tasks and determine that they are cached. After this change, it only takes 1.629 seconds (54% of what it used to be).

The optimizations are:

1. Don't create a new background thread to animate each spinner. Instead, create one background thread that acts as a "spinner service" and lives for the duration of the whole program. The cost of creating a new thread only takes 10 us or so, so this isn't a big speedup.
2. For the first 100ms of the spinner animation, just animate faster! The spinner takes up to a whole animation cycle to shut down. So for the first 100ms, we reduce the animation cycle from 100ms down to 16ms. Previously, running 30 no-ops would take up to 30 * 100ms = 3 seconds just for waiting on the spinner! Now the total spinner time for 30 no-ops is at most 480ms.

@juliahw